pyproject.toml
[project]
name = "d20"
version = "1.0.0"
description = "The basic blog app built in the Flask tutorial."
dependencies = [
    "flask",
    "plox",
    "Bootstrap-Flask",
    "pytest",
    "psycopg2-binary",
    "python-dotenv",
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"

[tool.pytest.ini_options]
testpaths = ["test"]

[tool.coverage.run]
branch = true
source = ["d20"]

